Refactor: Move unit registry to units.py
Move the unit registry to a dedicated units.py file to improve
organization and avoid circular dependencies.
